LATEST

+++ Model/bankaccount.cpp
@@ -1,6 +1,6 @@
 /**************************************************************************
 **   File: bankaccount.cpp
-**   Created on: Sat Oct 18 13:10:05 EET 2014
+**   Created on: Tue Nov 11 17:36:07 EET 2014
 **   Author: Michael Bishara
 **   Copyright: SphinxSolutions.
 **************************************************************************/
@@ -8,7 +8,7 @@
 #include "erpmodel.h"
 
 BankAccount::BankAccount()
-{
+ : QSqlRelationalTableModel(){
 
 this->BankAccountID = 0 ;
 this->Name = " ";
@@ -23,8 +23,13 @@ this->CurrencyID = 0 ;
 this->ContactID = 0 ;
 this->CountryID = 0 ;
 this->BankCountryCode = " ";
+this->setTable("BankAccount");
+this->setEditStrategy(QSqlTableModel::OnManualSubmit);
+this->setRelation(9, QSqlRelation("Currency", "CurrencyID", "Description"));
+this->setRelation(10, QSqlRelation("Contact", "ContactID", "Name"));
+this->setRelation(11, QSqlRelation("Country", "CountryID", "Name"));
 }
-BankAccount::BankAccount(int BankAccountID,QString Name,QString BankAddress,QString BankCode,QString AccountName,QString AccountNumber,QString IBAN,QString BIC,QString ZipCode,int CurrencyID,int ContactID,int CountryID,QString BankCountryCode){
+BankAccount::BankAccount(int BankAccountID,QString Name,QString BankAddress,QString BankCode,QString AccountName,QString AccountNumber,QString IBAN,QString BIC,QString ZipCode,int CurrencyID,int ContactID,int CountryID,QString BankCountryCode) : QSqlRelationalTableModel(){
 this->BankAccountID = BankAccountID ;
 this->Name = Name ;
 this->BankAddress = BankAddress ;
@@ -40,7 +45,8 @@ this->CountryID = CountryID ;
 this->BankCountryCode = BankCountryCode ;
 }
 
-BankAccount::BankAccount(QString Name,QString BankAddress,QString BankCode,QString AccountName,QString AccountNumber,QString IBAN,QString BIC,QString ZipCode,int CurrencyID,int ContactID,int CountryID,QString BankCountryCode){
+BankAccount::BankAccount(QString Name,QString BankAddress,QString BankCode,QString AccountName,QString AccountNumber,QString IBAN,QString BIC,QString ZipCode,int CurrencyID,int ContactID,int CountryID,QString BankCountryCode) : QSqlRelationalTableModel(){
+this->BankAccountID = 0 ;
 this->Name = Name ;
 this->BankAddress = BankAddress ;
 this->BankCode = BankCode ;
@@ -101,8 +107,7 @@ return true;
 
 bool BankAccount::remove() {
 if(BankAccountID!= 0) {
-(ErpModel::GetInstance()->qeryExec("DELETE FROM BankAccount"
-"WHERE BankAccountID ='"+QString::number(this->BankAccountID)+"'"));
+(ErpModel::GetInstance()->qeryExec("DELETE FROM BankAccount WHERE BankAccountID ="+QString::number(this->BankAccountID)+""));
 return true;
  }
 return false;
@@ -143,8 +148,7 @@ return new BankAccount();
 
 BankAccount* BankAccount::get(QString name) {
 if(name != NULL) {
-QSqlQuery query = (ErpModel::GetInstance()->qeryExec("SELECT *  FROM BankAccount"
-"WHERE Name = '"+name+"'"));
+QSqlQuery query = (ErpModel::GetInstance()->qeryExec("SELECT *  FROM BankAccount WHERE Name = '"+name+"'"));
 while (query.next()) {
 return new BankAccount(query.value(0).toInt(),query.value(1).toString(),query.value(2).toString(),query.value(3).toString(),query.value(4).toString(),query.value(5).toString(),query.value(6).toString(),query.value(7).toString(),query.value(8).toString(),query.value(9).toInt(),query.value(10).toInt(),query.value(11).toInt(),query.value(12).toString());
  }
@@ -188,6 +192,18 @@ QList<QString> BankAccount::GetStringList() {
 	return list;
 }
 
+QHash<int,QString> BankAccount::GetHashList() {
+	QHash<int,QString> list;
+	int count =BankAccount::GetInstance()->bankaccounts.count();
+	if( count < 2){
+		BankAccount::getAll();
+	}
+	for(int i = 0; i < count; i++){
+		list.insert(BankAccount::GetInstance()->bankaccounts[i]->BankAccountID,BankAccount::GetInstance()->bankaccounts[i]->Name);
+	}
+	return list;
+}
+
 int BankAccount::GetIndex(QString name) {
 	int count =BankAccount::GetInstance()->bankaccounts.count();
 	if( count < 2){
@@ -201,7 +217,7 @@ int BankAccount::GetIndex(QString name) {
 	return 0;
 }
 
-QList<BankAccount*> BankAccount::select(QString select) {
+QList<BankAccount*> BankAccount::querySelect(QString select) {
 QList<BankAccount*>list;
 if(select != NULL) {
 QSqlQuery query = (ErpModel::GetInstance()->qeryExec("SELECT *  FROM BankAccount"
@@ -213,3 +229,186 @@ list.append(new BankAccount(query.value(0).toInt(),query.value(1).toString(),que
 }
 return list;
  }
+
+Qt::ItemFlags BankAccount::flags(const QModelIndex &index) const {
+Qt::ItemFlags flags = QSqlRelationalTableModel::flags(index);
+flags ^= Qt::ItemIsEditable;
+if (
+index.column() == 1 || index.column() == 2 || index.column() == 3 || index.column() == 4 || index.column() == 5 || index.column() == 6 || index.column() == 7 || index.column() == 8 || index.column() == 9 || index.column() == 10 || index.column() == 11 || index.column() == 12)
+flags |= Qt::ItemIsEditable;
+return flags;
+}
+
+bool BankAccount::setData(const QModelIndex &index, const QVariant &value, int /* role */) {
+	if (index.column() < 1)
+		return false;
+QModelIndex primaryKeyIndex = QSqlRelationalTableModel::index(index.row(), 0);
+int id = data(primaryKeyIndex).toInt();
+bool ok = true;
+if((data(QSqlRelationalTableModel::index(index.row(), index.column())).toString() != value.toString().toLower())){
+if (index.column() == 1)
+ok = setName(id, value.toString());
+else if (index.column() == 2)
+ok = setBankAddress(id, value.toString());
+else if (index.column() == 3)
+ok = setBankCode(id, value.toString());
+else if (index.column() == 4)
+ok = setAccountName(id, value.toString());
+else if (index.column() == 5)
+ok = setAccountNumber(id, value.toString());
+else if (index.column() == 6)
+ok = setIBAN(id, value.toString());
+else if (index.column() == 7)
+ok = setBIC(id, value.toString());
+else if (index.column() == 8)
+ok = setZipCode(id, value.toString());
+else if (index.column() == 9)
+ok = setCurrencyID(id, value.toString());
+else if (index.column() == 10)
+ok = setContactID(id, value.toString());
+else if (index.column() == 11)
+ok = setCountryID(id, value.toString());
+else if (index.column() == 12)
+ok = setBankCountryCode(id, value.toString());
+refresh();
+}
+return ok;
+}
+
+bool BankAccount::remove(const QModelIndex &index) {
+QModelIndex primaryKeyIndex = QSqlRelationalTableModel::index(index.row(), 0);
+bool ok = true;
+this->BankAccountID = data(primaryKeyIndex).toInt();
+ok = this->remove();
+refresh();
+return ok;
+}
+
+void BankAccount::refresh() {
+if(!ErpModel::GetInstance()->db.isOpen()&&!ErpModel::GetInstance()->db.open())
+qDebug() <<"Couldn't open DataBase at Refresh() BankAccount!";
+this->setHeaderData(1, Qt::Horizontal, QObject::tr("Name"));
+this->setHeaderData(2, Qt::Horizontal, QObject::tr("Bank Address"));
+this->setHeaderData(3, Qt::Horizontal, QObject::tr("Bank Code"));
+this->setHeaderData(4, Qt::Horizontal, QObject::tr("Account Name"));
+this->setHeaderData(5, Qt::Horizontal, QObject::tr("Account Number"));
+this->setHeaderData(6, Qt::Horizontal, QObject::tr("I B A N"));
+this->setHeaderData(7, Qt::Horizontal, QObject::tr("B I C"));
+this->setHeaderData(8, Qt::Horizontal, QObject::tr("Zip Code"));
+this->setHeaderData(9, Qt::Horizontal, QObject::tr("Currency"));
+this->setHeaderData(10, Qt::Horizontal, QObject::tr("Contact"));
+this->setHeaderData(11, Qt::Horizontal, QObject::tr("Country"));
+this->setHeaderData(12, Qt::Horizontal, QObject::tr("Bank Country Code"));
+	this->select();
+//	if(ErpModel::GetInstance()->db.isOpen())
+//		ErpModel::GetInstance()->db.close();
+}
+bool BankAccount::setName(int BankAccountID, const QString &Name) {
+QSqlQuery query;
+query.prepare("update BankAccount set Name = ? where BankAccountID = ?");
+query.addBindValue(Name);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setBankAddress(int BankAccountID, const QString &BankAddress) {
+QSqlQuery query;
+query.prepare("update BankAccount set BankAddress = ? where BankAccountID = ?");
+query.addBindValue(BankAddress);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setBankCode(int BankAccountID, const QString &BankCode) {
+QSqlQuery query;
+query.prepare("update BankAccount set BankCode = ? where BankAccountID = ?");
+query.addBindValue(BankCode);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setAccountName(int BankAccountID, const QString &AccountName) {
+QSqlQuery query;
+query.prepare("update BankAccount set AccountName = ? where BankAccountID = ?");
+query.addBindValue(AccountName);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setAccountNumber(int BankAccountID, const QString &AccountNumber) {
+QSqlQuery query;
+query.prepare("update BankAccount set AccountNumber = ? where BankAccountID = ?");
+query.addBindValue(AccountNumber);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setIBAN(int BankAccountID, const QString &IBAN) {
+QSqlQuery query;
+query.prepare("update BankAccount set IBAN = ? where BankAccountID = ?");
+query.addBindValue(IBAN);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setBIC(int BankAccountID, const QString &BIC) {
+QSqlQuery query;
+query.prepare("update BankAccount set BIC = ? where BankAccountID = ?");
+query.addBindValue(BIC);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setZipCode(int BankAccountID, const QString &ZipCode) {
+QSqlQuery query;
+query.prepare("update BankAccount set ZipCode = ? where BankAccountID = ?");
+query.addBindValue(ZipCode);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setCurrencyID(int BankAccountID, const QString &CurrencyID) {
+QSqlQuery query;
+query.prepare("update BankAccount set CurrencyID = ? where BankAccountID = ?");
+query.addBindValue(CurrencyID);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setContactID(int BankAccountID, const QString &ContactID) {
+QSqlQuery query;
+query.prepare("update BankAccount set ContactID = ? where BankAccountID = ?");
+query.addBindValue(ContactID);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setCountryID(int BankAccountID, const QString &CountryID) {
+QSqlQuery query;
+query.prepare("update BankAccount set CountryID = ? where BankAccountID = ?");
+query.addBindValue(CountryID);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool BankAccount::setBankCountryCode(int BankAccountID, const QString &BankCountryCode) {
+QSqlQuery query;
+query.prepare("update BankAccount set BankCountryCode = ? where BankAccountID = ?");
+query.addBindValue(BankCountryCode);
+query.addBindValue(BankAccountID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+
