+++ Model/user.cpp
@@ -1,6 +1,6 @@
 /**************************************************************************
 **   File: user.cpp
-**   Created on: Sat Oct 18 13:10:05 EET 2014
+**   Created on: Tue Nov 11 17:36:07 EET 2014
 **   Author: Michael Bishara
 **   Copyright: SphinxSolutions.
 **************************************************************************/
@@ -8,7 +8,7 @@
 #include "erpmodel.h"
 
 User::User()
-{
+ : QSqlRelationalTableModel(){
 
 this->UserID = 0 ;
 this->Name = " ";
@@ -18,8 +18,11 @@ this->LastLogin = " ";
 this->EmployeeID = 0 ;
 this->active = 0 ;
 this->lastIP = " ";
+this->setTable("User");
+this->setEditStrategy(QSqlTableModel::OnManualSubmit);
+this->setRelation(5, QSqlRelation("Employee", "EmployeeID", "Name"));
 }
-User::User(int UserID,QString Name,QString Username,QString Password,QString LastLogin,int EmployeeID,bool active,QString lastIP){
+User::User(int UserID,QString Name,QString Username,QString Password,QString LastLogin,int EmployeeID,bool active,QString lastIP) : QSqlRelationalTableModel(){
 this->UserID = UserID ;
 this->Name = Name ;
 this->Username = Username ;
@@ -30,7 +33,8 @@ this->active = active ;
 this->lastIP = lastIP ;
 }
 
-User::User(QString Name,QString Username,QString Password,QString LastLogin,int EmployeeID,bool active,QString lastIP){
+User::User(QString Name,QString Username,QString Password,QString LastLogin,int EmployeeID,bool active,QString lastIP) : QSqlRelationalTableModel(){
+this->UserID = 0 ;
 this->Name = Name ;
 this->Username = Username ;
 this->Password = Password ;
@@ -79,8 +83,7 @@ return true;
 
 bool User::remove() {
 if(UserID!= 0) {
-(ErpModel::GetInstance()->qeryExec("DELETE FROM User"
-"WHERE UserID ='"+QString::number(this->UserID)+"'"));
+(ErpModel::GetInstance()->qeryExec("DELETE FROM User WHERE UserID ="+QString::number(this->UserID)+""));
 return true;
  }
 return false;
@@ -121,8 +124,7 @@ return new User();
 
 User* User::get(QString name) {
 if(name != NULL) {
-QSqlQuery query = (ErpModel::GetInstance()->qeryExec("SELECT *  FROM User"
-"WHERE Name = '"+name+"'"));
+QSqlQuery query = (ErpModel::GetInstance()->qeryExec("SELECT *  FROM User WHERE Name = '"+name+"'"));
 while (query.next()) {
 return new User(query.value(0).toInt(),query.value(1).toString(),query.value(2).toString(),query.value(3).toString(),query.value(4).toString(),query.value(5).toInt(),query.value(6).toInt(),query.value(7).toString());
  }
@@ -162,6 +164,18 @@ QList<QString> User::GetStringList() {
 	return list;
 }
 
+QHash<int,QString> User::GetHashList() {
+	QHash<int,QString> list;
+	int count =User::GetInstance()->users.count();
+	if( count < 2){
+		User::getAll();
+	}
+	for(int i = 0; i < count; i++){
+		list.insert(User::GetInstance()->users[i]->UserID,User::GetInstance()->users[i]->Name);
+	}
+	return list;
+}
+
 int User::GetIndex(QString name) {
 	int count =User::GetInstance()->users.count();
 	if( count < 2){
@@ -175,7 +189,7 @@ int User::GetIndex(QString name) {
 	return 0;
 }
 
-QList<User*> User::select(QString select) {
+QList<User*> User::querySelect(QString select) {
 QList<User*>list;
 if(select != NULL) {
 QSqlQuery query = (ErpModel::GetInstance()->qeryExec("SELECT *  FROM User"
@@ -187,3 +201,126 @@ list.append(new User(query.value(0).toInt(),query.value(1).toString(),query.valu
 }
 return list;
  }
+
+Qt::ItemFlags User::flags(const QModelIndex &index) const {
+Qt::ItemFlags flags = QSqlRelationalTableModel::flags(index);
+flags ^= Qt::ItemIsEditable;
+if (
+index.column() == 1 || index.column() == 2 || index.column() == 3 || index.column() == 4 || index.column() == 5 || index.column() == 6 || index.column() == 7)
+flags |= Qt::ItemIsEditable;
+return flags;
+}
+
+bool User::setData(const QModelIndex &index, const QVariant &value, int /* role */) {
+	if (index.column() < 1)
+		return false;
+QModelIndex primaryKeyIndex = QSqlRelationalTableModel::index(index.row(), 0);
+int id = data(primaryKeyIndex).toInt();
+bool ok = true;
+if((data(QSqlRelationalTableModel::index(index.row(), index.column())).toString() != value.toString().toLower())){
+if (index.column() == 1)
+ok = setName(id, value.toString());
+else if (index.column() == 2)
+ok = setUsername(id, value.toString());
+else if (index.column() == 3)
+ok = setPassword(id, value.toString());
+else if (index.column() == 4)
+ok = setLastLogin(id, value.toString());
+else if (index.column() == 5)
+ok = setEmployeeID(id, value.toString());
+else if (index.column() == 6)
+ok = setactive(id, value.toString());
+else if (index.column() == 7)
+ok = setlastIP(id, value.toString());
+refresh();
+}
+return ok;
+}
+
+bool User::remove(const QModelIndex &index) {
+QModelIndex primaryKeyIndex = QSqlRelationalTableModel::index(index.row(), 0);
+bool ok = true;
+this->UserID = data(primaryKeyIndex).toInt();
+ok = this->remove();
+refresh();
+return ok;
+}
+
+void User::refresh() {
+if(!ErpModel::GetInstance()->db.isOpen()&&!ErpModel::GetInstance()->db.open())
+qDebug() <<"Couldn't open DataBase at Refresh() User!";
+this->setHeaderData(1, Qt::Horizontal, QObject::tr("Name"));
+this->setHeaderData(2, Qt::Horizontal, QObject::tr("Username"));
+this->setHeaderData(3, Qt::Horizontal, QObject::tr("Password"));
+this->setHeaderData(4, Qt::Horizontal, QObject::tr("Last Login"));
+this->setHeaderData(5, Qt::Horizontal, QObject::tr("Employee"));
+this->setHeaderData(6, Qt::Horizontal, QObject::tr("active"));
+this->setHeaderData(7, Qt::Horizontal, QObject::tr("lastI P"));
+	this->select();
+//	if(ErpModel::GetInstance()->db.isOpen())
+//		ErpModel::GetInstance()->db.close();
+}
+bool User::setName(int UserID, const QString &Name) {
+QSqlQuery query;
+query.prepare("update User set Name = ? where UserID = ?");
+query.addBindValue(Name);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool User::setUsername(int UserID, const QString &Username) {
+QSqlQuery query;
+query.prepare("update User set Username = ? where UserID = ?");
+query.addBindValue(Username);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool User::setPassword(int UserID, const QString &Password) {
+QSqlQuery query;
+query.prepare("update User set Password = ? where UserID = ?");
+query.addBindValue(Password);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool User::setLastLogin(int UserID, const QString &LastLogin) {
+QSqlQuery query;
+query.prepare("update User set LastLogin = ? where UserID = ?");
+query.addBindValue(LastLogin);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool User::setEmployeeID(int UserID, const QString &EmployeeID) {
+QSqlQuery query;
+query.prepare("update User set EmployeeID = ? where UserID = ?");
+query.addBindValue(EmployeeID);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool User::setactive(int UserID, const QString &active) {
+QSqlQuery query;
+query.prepare("update User set active = ? where UserID = ?");
+query.addBindValue(active);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+bool User::setlastIP(int UserID, const QString &lastIP) {
+QSqlQuery query;
+query.prepare("update User set lastIP = ? where UserID = ?");
+query.addBindValue(lastIP);
+query.addBindValue(UserID);
+if( !query.exec() )
+qDebug() << query.lastError().text();
+return true;
+}
+
